+++ test/unit/rx-database.test.ts
@@ -183,7 +183,7 @@ config.parallel('rx-database.test.js', () => {
                     password,
                     ignoreDuplicate: true
                 });
-                const doc = await db._adminPouch.get('_local/pwHash');
+                const doc = await db.internalStore.get('_local/pwHash');
                 assert.strictEqual(typeof doc.value, 'string');
                 const db2 = await createRxDatabase({
                     name,
@@ -191,7 +191,7 @@ config.parallel('rx-database.test.js', () => {
                     password,
                     ignoreDuplicate: true
                 });
-                const doc2 = await db._adminPouch.get('_local/pwHash');
+                const doc2 = await db.internalStore.get('_local/pwHash');
                 assert.ok(doc2);
                 assert.strictEqual(typeof doc.value, 'string');
 
@@ -260,7 +260,7 @@ config.parallel('rx-database.test.js', () => {
                     name: 'human0',
                     schema: schemas.human
                 });
-                const colDoc = await db._collectionsPouch.get('human0-' + schemas.human.version);
+                const colDoc = await db.internalStore.get('human0-' + schemas.human.version);
                 const compareSchema = createRxSchema(schemas.human);
                 assert.deepStrictEqual(compareSchema.normalized, colDoc.schema);
                 db.destroy();
@@ -289,7 +289,7 @@ config.parallel('rx-database.test.js', () => {
                 });
                 const version = collection.schema.version;
                 assert.deepStrictEqual(version, 0);
-                const internalDoc = await db._collectionsPouch.get('human-' + version);
+                const internalDoc = await db.internalStore.get('human-' + version);
                 assert.deepStrictEqual(internalDoc.version, version);
                 db.destroy();
             });