+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Breaking:
   - The internal `hash()` function does now use a RxDB specific salt
   - Renamed `RxSchema.jsonID` to `RxSchema.jsonSchema`
   - Moved remaining stuff of leader-election from core into the plugin
+  - Merged multiple internal databases for metadata into one `internalStore`
 
 Other:
   - Removed many runtime type checks that now should be covered by typescript in buildtime